move product render to board edit form
