+++ src/modules/boards/components/editForm/EditForm.tsx
@@ -1,6 +1,10 @@
+import { HeaderContentSmall } from 'modules/boards/styles/item';
+import ControlLabel from 'modules/common/components/form/Label';
 import Icon from 'modules/common/components/Icon';
 import { CloseModal } from 'modules/common/styles/main';
 import { __, extractAttachment } from 'modules/common/utils';
+import ProductSection from 'modules/deals/components/ProductSection';
+import { IProduct } from 'modules/settings/productService/types';
 import React from 'react';
 import Modal from 'react-bootstrap/Modal';
 import { IEditFormContent, IItem, IItemParams, IOptions } from '../../types';
@@ -23,16 +27,103 @@ type State = {
   stageId?: string;
   updatedItem?: IItem;
   prevStageId?: string;
+  amount?: any;
+  products?: IProduct[];
+  productsData?: any;
 };
 
 class EditForm extends React.Component<Props, State> {
   constructor(props) {
     super(props);
 
+    const item = props.item;
+
     this.state = {
-      stageId: props.item.stageId
+      stageId: item.stageId,
+      amount: item.amount || {},
+      productsData: item.products ? item.products.map(p => ({ ...p })) : [],
+      // collecting data for ItemCounter component
+      products: item.products ? item.products.map(p => p.product) : []
     };
   }
+
+  renderAmount = () => {
+    const { amount } = this.state;
+
+    if (Object.keys(amount).length === 0) {
+      return null;
+    }
+
+    return (
+      <HeaderContentSmall>
+        <ControlLabel>{__('Amount')}</ControlLabel>
+        {Object.keys(amount).map(key => (
+          <p key={key}>
+            {amount[key].toLocaleString()} {key}
+          </p>
+        ))}
+      </HeaderContentSmall>
+    );
+  };
+
+  saveProductsData = () => {
+    const { productsData } = this.state;
+    const products: IProduct[] = [];
+    const amount: any = {};
+
+    const filteredProductsData: any = [];
+
+    productsData.forEach(data => {
+      // products
+      if (data.product) {
+        if (data.currency) {
+          // calculating item amount
+          if (!amount[data.currency]) {
+            amount[data.currency] = data.amount || 0;
+          } else {
+            amount[data.currency] += data.amount || 0;
+          }
+        }
+
+        // collecting data for ItemCounter component
+        products.push(data.product);
+
+        data.productId = data.product._id;
+
+        filteredProductsData.push(data);
+      }
+    });
+
+    this.setState({ productsData: filteredProductsData, products }, () => {
+      this.props.saveItem(
+        { productsData: this.state.productsData },
+        updatedItem => {
+          this.props.onUpdate(updatedItem);
+        }
+      );
+    });
+  };
+
+  onChangeField = <T extends keyof State>(name: T, value: State[T]) => {
+    this.setState({ [name]: value } as Pick<State, keyof State>);
+  };
+
+  renderProductSection = () => {
+    const { products, productsData } = this.state;
+
+    const pDataChange = pData => this.onChangeField('productsData', pData);
+    const prsChange = prs => this.onChangeField('products', prs);
+
+    return (
+      <ProductSection
+        onChangeProductsData={pDataChange}
+        onChangeProducts={prsChange}
+        productsData={productsData}
+        products={products || []}
+        saveProductsData={this.saveProductsData}
+      />
+    );
+  };
 
   onChangeStage = (stageId: string) => {
     this.setState({ stageId }, () => {
@@ -129,7 +220,9 @@ class EditForm extends React.Component<Props, State> {
             saveItem: this.saveItem,
             onChangeStage: this.onChangeStage,
             copy: this.copy,
-            remove: this.remove
+            remove: this.remove,
+            renderAmount: this.renderAmount,
+            renderSidebar: this.renderProductSection
           })}
         </Modal.Body>
       </Modal>
